+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [v0.308.0] - 2026-02-17
 
+### Added
+
+- `pt-saft-v1`: Added correction definitions
+
 ### Removed
 
 - `pkg/template`: removed as no longer used.